Refine rollout crate extraction
