+++ codex-rs/core/src/lib.rs
@@ -68,6 +68,7 @@ mod model_provider_info;
 pub mod path_utils;
 pub mod personality_migration;
 pub mod plugins;
+mod rollout_config_builder;
 mod sandbox_tags;
 pub mod sandboxing;
 mod session_prefix;
@@ -128,7 +129,7 @@ pub mod shell_snapshot;
 pub mod skills;
 pub mod spawn;
 mod state;
-pub mod state_db;
+pub mod state_runtime;
 mod tasks;
 mod tools;
 pub mod turn_diff_tracker;
@@ -157,6 +158,8 @@ pub use exec_policy::check_execpolicy_for_warnings;
 pub use exec_policy::format_exec_policy_error_with_source;
 pub use exec_policy::load_exec_policy;
 pub use file_watcher::FileWatcherEvent;
+pub use rollout_config_builder::rollout_config;
+pub use rollout_config_builder::rollout_config_with_cwd;
 pub use safety::get_platform_sandbox;
 pub use tools::spec::parse_tool_input_schema;
 pub use turn_metadata::build_turn_metadata_header;
