+++ codex-rs/rollout/src/lib.rs
@@ -20,6 +20,8 @@ pub mod recorder;
 pub mod session_index;
 mod state_db;
 #[cfg(test)]
+mod test_support;
+#[cfg(test)]
 mod tests;
 
 pub use list::Cursor;
