+++ codex-rs/rollout/src/state_db_tests.rs
@@ -1,6 +1,11 @@
 use super::*;
-use codex_rollout::parse_cursor;
+use crate::parse_cursor;
+use chrono::DateTime;
+use chrono::NaiveDateTime;
+use chrono::Timelike;
+use chrono::Utc;
 use pretty_assertions::assert_eq;
+use uuid::Uuid;
 
 #[test]
 fn cursor_to_anchor_normalizes_timestamp_format() {
