+++ codex-rs/tui_app_server/src/resume_picker.rs
@@ -19,11 +19,11 @@ use codex_app_server_protocol::ThreadSortKey as AppServerThreadSortKey;
 use codex_app_server_protocol::ThreadSourceKind;
 use codex_core::config::Config;
 use codex_core::path_utils;
-use codex_core::state_db::get_state_db;
+use codex_core::rollout_config;
+use codex_core::state_runtime::get_state_db;
 use codex_protocol::ThreadId;
 use codex_rollout::Cursor;
 use codex_rollout::INTERACTIVE_SESSION_SOURCES;
-use codex_rollout::RolloutConfig;
 use codex_rollout::RolloutRecorder;
 use codex_rollout::ThreadItem;
 use codex_rollout::ThreadSortKey;
@@ -117,16 +117,6 @@ enum ProviderFilter {
 }
 
 type PageLoader = Arc<dyn Fn(PageLoadRequest) + Send + Sync>;
-
-fn rollout_config(config: &Config) -> RolloutConfig {
-    RolloutConfig::new(
-        config.codex_home.clone(),
-        config.sqlite_home.clone(),
-        config.cwd.clone(),
-        config.model_provider_id.clone(),
-        config.memories.generate_memories,
-    )
-}
 
 enum BackgroundEvent {
     PageLoaded {
